Time: 0 ms (100.00%) | Memory: 6.4 MB (63.74%) - LeetSync

+++ 806-domino-and-tromino-tiling/domino-and-tromino-tiling.cpp
@@ -1,14 +1,13 @@
 class Solution {
 public:
     const int mod = 1e9+7;
-    long long dp[1000];
-    long long solve(int n )
+    int dp[1000];
+    int solve(int n )
     {
-        if(n<=2)return n ;
+        if(n<=2)
+            return n ;
         if(n==3)return 5;
-        
         if(dp[n]!=-1)return dp[n];
-       
         return dp[n] = (2*solve(n-1)%mod+solve(n-3)%mod)%mod;
     }
     int numTilings(int n) {
